refactor: move delegation migration test
